requirements.txt
Django==4.0.3
flake8==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.0.1
django-imagekit==4.1.0
geopy==2.2.0
django-filter==21.1
